# openapi/openapi.yaml
openapi: "3.0.0"
info:
  version: 0.1.5
  title: CRM API
  description: A main CRM Api
  contact:
    name: API Team
servers:
  - url: "localhost:8080"
    description: Dev server
...
paths:
  /legal-entities:
    get:
      tags:
        - federation
      summary: Get all legal entities
      operationId: getAllLegalEntities
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LegalEntityDTO'
    post:
      tags:
        - federation
      summary: Create new legal entity
      operationId: createLegalEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLegalEntityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalEntityDTO'
  /legal-entities/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags:
        - federation
      summary: Delete a legal entity
      operationId: deleteLegalEntity
      responses:
        '204':
          description: No Content
    put:
      tags:
        - federation
      summary: Update a legal entity
      operationId: updateLegalEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLegalEntityRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalEntityDTO'
...
  /bank-accounts/{legal_entity_uuid}:
    get:
      tags:
        - federation
      summary: Get all bank accounts for a specific legal entity
      operationId: getBankAccountsByLegalEntityUUID
      parameters:
        - name: legal_entity_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankAccountDTO'

    post:
      tags:
        - federation
      summary: Create a new bank account for a legal entity
      operationId: createBankAccount
      parameters:
        - name: legal_entity_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBankAccountRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountDTO'
  /bank-accounts/{uuid}:
    get:
      tags:
        - federation
      summary: Get a bank account by its UUID
      operationId: getBankAccount
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountDTO'

    put:
      tags:
        - federation
      summary: Update a bank account
      operationId: updateBankAccount
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBankAccountRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountDTO'

    delete:
      tags:
        - federation
      summary: Delete a bank account
      operationId: deleteBankAccount
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
...
  /profile/photo:
    patch:
      tags:
        - profile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Upload user photo to s3
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/ProfilePhotoDTO"

    delete:
      description: Delete user photo from s3
      tags:
        - profile
      responses:
        200:
          description: Ok
        401:
          description: Unauthorized
...
components:
...
  schemas:
    LegalEntityDTO:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
    CreateLegalEntityRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateLegalEntityRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    BankAccountDTO:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        bic:
          type: string
        bank_name:
          type: string
        address:
          type: string
        correspondent_account:
          type: string
        account_number:
          type: string
        currency:
          type: string
        comment:
          type: string
        legal_entity_uuid:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
    CreateBankAccountRequest:
      type: object
      properties:
        bic:
          type: string
        bank_name:
          type: string
        address:
          type: string
        correspondent_account:
          type: string
        account_number:
          type: string
        currency:
          type: string
        comment:
          type: string
      required:
        - bic
        - bank_name
        - account_number
    UpdateBankAccountRequest:
      type: object
      properties:
        bic:
          type: string
        bank_name:
          type: string
        address:
          type: string
        correspondent_account:
          type: string
        account_number:
          type: string
        currency:
          type: string
        comment:
          type: string
      required:
        - bic
        - bank_name
        - account_number
...
    ProfilePhotoDTO:
      x-go-type: dto.ProfilePhotoDTO
      x-go-type-import:
        name: ProfilePhotoDTOSurveyDTO
        path: github.com/krisch/crm-backend/dto
      type: object
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []